requirements.txt
pefile==2024.8.26
firebase-admin==7.1.0
python-dotenv